requirements.txt
pandas>=1.3.0
numpy>=1.21.0
torch>=2.0.0
diffusers>=0.27.0
transformers>=4.38.0
Pillow>=9.0.0
matplotlib>=3.4.0
opencv-python>=4.5.0
streamlit>=1.20.0
